+++ {{cookiecutter.project_slug}}/backend/config/settings/local.py
@@ -13,3 +13,13 @@ ALLOWED_HOSTS = [
     'django'
 ]
 # =============================================================
+
+{%- if cookiecutter.use_mailhog == "y" %}
+
+
+# EMAIL
+EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
+EMAIL_HOST = 'mailhog'
+EMAIL_PORT = 1025
+# =============================================================
+{% endif %}